+++ src/main/java/ru/aston/mcs/controller/ManagerController.java
@@ -3,10 +3,7 @@ package ru.aston.mcs.controller;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import ru.aston.mcs.dto.ManagerDTO;
-import ru.aston.mcs.entity.Manager;
 import ru.aston.mcs.service.ManagerService;
-
-import java.util.List;
 
 @RestController
 @RequestMapping("/api/managers")
@@ -16,17 +13,17 @@ public class ManagerController {
     private ManagerService managerService;
 
     @GetMapping("/{id}")
-    public ManagerDTO getManager(@PathVariable int id){
+    public ManagerDTO getManager(@PathVariable Long id) {
         return managerService.getManager(id);
     }
 
     @PostMapping("/")
-    public void addNewManager(@RequestBody ManagerDTO managerDTO){
+    public void addNewManager(@RequestBody ManagerDTO managerDTO) {
         managerService.saveManager(managerDTO);
     }
 
     @DeleteMapping("/{id}")
-    public void deleteManager(@PathVariable int id){
+    public void deleteManager(@PathVariable Long id) {
         managerService.deleteManager(id);
     }
 }
