+++ src/main/java/ru/aston/mcs/repository/ManagerRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import ru.aston.mcs.entity.Manager;
 
 @Repository
-public interface ManagerRepository extends JpaRepository<Manager, Integer> {
+public interface ManagerRepository extends JpaRepository<Manager, Long> {
 }
